requirements.txt
# Core AI / RAG
sentence-transformers>=3.0.0
torch>=2.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
rank-bm25>=0.2.2

# PostgreSQL + pgvector
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
